+++ CHANGELOG.md
@@ -6,6 +6,35 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
+<<<<<<< Updated upstream
+=======
+## 2025.11.07
+
+### Added
+
+-
+
+### Updated
+
+- Removed LogicGatesProof.java.
+- Changed LogicGates1 to LogicGatesTree.
+
+
+## 2025.11.07
+
+### Added
+
+- Implemented secondary methods in LogicGatesSecondary.java.
+- Added new method in LogicGates.java to allow the change of inputs, and not just Operators.
+- Added LogicGates1.java file to call constructor in LogicGatesSecondary.java. As of the current moment, it contains no proper implementations.
+
+### Updated
+
+- Changed design to implement enum operators.
+- Changed design to make inputs for disassembling and assembling to LogicGates components. This is to better mimic binary tree behavior.
+- Moved height method out of secondary and into kernel.
+
+>>>>>>> Stashed changes
 ## 2025.10.18
 
 ### Added